+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
+## [1.0.2] - 2019-01-03
+### Changed
+- Fixes a bug that occurs when reconnecting.
+
 ## [1.0.1] - 2018-12-31
-### Added
-- Inital release.
+### Changed
+- Updates face-command-client.
+
+## [1.0.1] - 2018-12-31
+### Changed
+- Updates face-command-client.
 
 ## [1.0.0] - 2018-12-31
 ### Added